+++ Simple-PHP-IPAM/demo_seed.php
@@ -23,7 +23,7 @@ if (!($config['demo_mode']['enabled'] ?? false)) {
     exit(1);
 }
 
-$db = ipam_db($config);
+$db = ipam_db($config['db_path']);
 ipam_db_init($db);
 
 echo "Resetting database to demo data...\n";